separate build directory
