+++ renderer.js
@@ -1,14 +1,5 @@
 // This file is required by the index.html file and will
 // be executed in the renderer process for that window.
 // All of the Node.js APIs are available in this process.
-const shrdhtml = require('./shrdlite-html.js')
-
-//  <script type="text/javascript" src="lib/nearley.js"></script>
-//  <script type="text/javascript" src="grammar.js"></script>
-
-//  <script type="text/javascript" src="Shrdlite.js"></script>
-//  <script type="text/javascript" src="SVGWorld.js"></script>
-
-//  <script type="text/javascript" src="shrdlite-html.js"></script>
-
+const shrdhtml = require('./generated_code/shrdlite-html.js')
 shrdhtml.set_current_world();
